% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ASV_name_to_ASV_results.R
\name{add_ASV_name_to_ASV_results}
\alias{add_ASV_name_to_ASV_results}
\title{Adds taxonomic info to a data.frame/tibble}
\usage{
add_ASV_name_to_ASV_results(df, col_ASV, physeq, tax_lvls = "all")
}
\arguments{
\item{df}{`data.frame` or `tibble` containing some taxonomic information}

\item{col_ASV}{ASV, or in general taxon, column in the `data`}

\item{physeq}{phyloseq object, containing a `tax_table` with a column matching `col_ASV` name}

\item{tax_lvls}{a character vector, specifying the column names to add from the taxonomic table. Default is "all", which includes all levels}
}
\value{
A data.frame with taxon identifications added
}
\description{
This funcion adds some levels of phylogeny based on a phyloseq object onto a `data.frame` or `tibble` containing basic taxonomic information in a column
}
\examples{
data(GlobalPatterns)
library(tidyverse)
taxon_prevalences <- microbiome::prevalence(GlobalPatterns) \%>\%
    as.data.frame() \%>\%
    rownames_to_column("OTU") \%>\% # create a column containing the unique taxon identifier
    set_names(c("OTU", "prevalences")) \%>\% # to make sure we have column names we want
    add_ASV_name_to_ASV_results(df = ., col_ASV = "OTU", physeq = GlobalPatterns, tax_lvls = c("Phylum", "Class", "Genus"))


}
