% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_percent_columns.R
\name{add_percent_columns}
\alias{add_percent_columns}
\title{Add percentages to multicolumn tables}
\usage{
add_percent_columns(tb)
}
\arguments{
\item{tb}{a `table` R object}
}
\value{
a `matrix` with percent columns added next to each table column
}
\description{
Takes a complex table object as input, and it calculates the percentage for each column
}
\examples{

data(mtcars)
tb0 <- table(mtcars$cyl, mtcars$gear)
tb1 <- add_percent_columns(tb = tb0)


}
