% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_div_betw_withn.R
\name{beta_div_betw_withn}
\alias{beta_div_betw_withn}
\title{Beta diversity into group boxplots}
\usage{
beta_div_betw_withn(
  physeq,
  dist,
  variab,
  verbose = FALSE,
  btwn_pairwise = FALSE
)
}
\arguments{
\item{physeq}{a \code{phyloseq} object, important for the metadata it contains}

\item{dist}{either specify a metric (see \code{?phyloseq::distance} for more info), or pass a distance matrix either in the \code{matrix} or \code{dist} format}

\item{variab}{variable to segregate beta diversity values by. must be a \code{character} with a name in the \code{physeq}'s metadata}

\item{verbose}{should it print out the steps it is performing? (default is \code{FALSE})}

\item{btwn_pairwise}{should a third element be calculated and appended to the results? these will be pairwise comparisons of samples in each variable level.}
}
\value{
a \code{data.frame} object containing within and between diversities based on the variable specified. "between_pairwise" are calculated as "for each level of the variable categories, find diversities of all samples belonging to it against all other samples not belonging to it".
}
\description{
Function taking a phyloseq object or a diversity matrix and spits out group-wise interindividual diversity
}
\examples{

library(phyloseq)
data(GlobalPatterns)

betadiversitiesFull <- beta_div_betw_withn(physeq = GlobalPatterns,
                                           dist = "bray",
                                           variab = "Site",
                                           verbose = TRUE,
                                           btwn_pairwise = TRUE)

betadiversitiesSimpler <- beta_div_betw_withn(physeq = GlobalPatterns,
                                              dist = "bray",
                                              variab = "Site",
                                              verbose = TRUE,
                                              btwn_pairwise = FALSE)
}
