% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_integer_into_integer.R
\name{bin_integer_into_integer}
\alias{bin_integer_into_integer}
\title{Bin an integer/numeric variable}
\usage{
bin_integer_into_integer(x, multiples, include.lowest = FALSE)
}
\arguments{
\item{multiples}{An `integer` or `numeric` vector to bin `x` with}

\item{A}{`numeric` vector}

\item{include_highest}{Do you want to include upper or lower values in the bin? eg: if `multiples = 5` include_highest --> 'value <= 5', otherwise 'value < 5'}
}
\value{

}
\description{
Numeric binning means that any value even closer to its next multiple, is still binned to its lowest. Example: with "multiple = 5", 4.9 is still inside the 5 bin, but 5.1 is in the 10 bin. this is not a good approximation, but sometimes you need such an approximation
}
\examples{
data(mtcars)
bin_integer_into_integer(mtcars$mpg, multiples = 10, include_highest= TRUE)
bin_integer_into_integer(mtcars$mpg, multiples = 5, include_highest= FALSE)

}
