% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone_dir_structure.R
\name{clone_dir_structure}
\alias{clone_dir_structure}
\title{Clone directory structure (Not working yet)}
\usage{
clone_dir_structure(start_dir, end_dir)
}
\arguments{
\item{start_dir}{a character vector with a path to an existing directory}

\item{end_dir}{a character vector with a path to an existing directory}
}
\description{
Takes a directory and clones its structure elsewhere, but without copying files
}
\examples{

dir.create("~/tmp")
dir.create("~/tmp/tmp2", recursive = TRUE)
dir.create("~/tmp/tmp3", recursive = TRUE)
dir.create("~/tmp/tmp2/tmp2.1", recursive = TRUE)
dir.create("~/tmp/tmp3/tmp3.1", recursive = TRUE)

dir.create("~/destination_dir")
clone_dir_structure(start_dir = "~/tmp/",
                    end_dir = "~/destination_dir/")

dir()


}
