% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_Season.R
\name{date_to_Season}
\alias{date_to_Season}
\title{From date to its season}
\usage{
date_to_Season(date)
}
\arguments{
\item{date}{a \\code{Date} object}
}
\value{
An ordered \\code{factor}
}
\description{
Convert an object of Date class and turn into an ordered season factor, with Spring<Summer<Autumn<Winter
}
\examples{
dat <- c("2022-09-28", "2021-07-15", "2020-02-10", "2019-04-19")
season <- date_to_Season(date)

levels(season)

# want to have an unordered factor?
season <- factor(season, ordered = FALSE)
# or, with magrittr
# season \%<>\% factor(ordered = FALSE)

}
