% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declutter.envir.R
\name{declutter.envir}
\alias{declutter.envir}
\title{Delete temporary or test objects in `ls()`}
\usage{
declutter.envir(extra_rules = "test", verbose = FALSE)
}
\arguments{
\item{extra_rules}{defines extra patterns to delete (eg: extra_rules = c("temp","delete"))}
}
\description{
This function deletes all items (object, function, ...) containing either 'tmp' or 'test' in the name. Also, if the name is only 1 letter long, it will be deleted
}
\examples{
declutter_Envir(extra_rules = c("papaya", "attempt"))
}
