% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_mean.R
\name{gm_mean}
\alias{gm_mean}
\title{Geometric Mean}
\usage{
gm_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{raw numeric values}

\item{na.rm}{should NA values be removed? (default: TRUE)}
}
\value{
a numeric vector containing transformed values
}
\description{
Geometric Mean
}
\examples{

gm_mean(gss_cat$age)

}
