% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.nucleotide.R
\name{is.nucleotide}
\alias{is.nucleotide}
\title{Is a sequence made of nucleotides?}
\usage{
is.nucleotide(x)
}
\arguments{
\item{x}{a character vector to test}
}
\value{
a logical, saying if the character is only mad of nucleotide-typic characters (capital and non-capital A, T, G, C, U, N). NB: it does not support other characters, like W or whatnot.
}
\description{
Is a sequence made of nucleotides?
}
\examples{

is.nucleotide("ATTGCATA")
is.nucleotide("aTTgNGnt")
is.nucleotide("boyscout")

}
