% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_demogr_table_INCOMPLETE.R
\name{make_demogr_table}
\alias{make_demogr_table}
\title{Make a complex demographic table}
\usage{
make_demogr_table(
  df,
  row_variables_as_are,
  row_var_names_nicer,
  col_variable,
  col_var_names_nicer,
  add_perc_column,
  caption = NULL,
  cont_vars_summary = "mean_sd"
)
}
\arguments{
\item{df}{A `data.frame` or `tibble` containing several variables of interest to make a table from}

\item{row_variables_as_are}{a `vector` of column names to subdivide rows by}

\item{row_var_names_nicer}{(Facultative) a character vector of variables names in `row_variables_as_are`, but more "presentation ready"}

\item{col_variable}{A `character` vector with the name of the df colum which will become columns of the}

\item{col_var_names_nicer}{if you want nicer-looking names to the column variable items/levels,}

\item{add_perc_column}{Add per-table-column's percentage column}

\item{caption}{. Default is `NULL`}

\item{cont_var_summary}{could be either "mean_sd" or "median_iqr"}
}
\value{
a KableExtra HTML table object
}
\description{
An attempt to make a nice and flexible demographic table, possibly also adding significance when applicable... too ambitious though, probably I will never make it work. so far it doesn't, indeed.
}
\details{
Summarize multiple **categorical** variable columns in one handy function
NB: Doesn't yet work perfectly, also, it doesn't make summary statistics of continuous variables
}
\examples{
make_demogr_table(df= mtcars,
                  row_variables_as_are = c("mpg","gear", "carb"),
                  col_var_names_nicer = c("4", "6", "8"),
                  row_var_names_nicer = c("Miles per gallon", "N? gears", "Fuel type"),
                  col_variable = "cyl",
                  add_perc_column = TRUE)
}
