% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy_add_metadata_variables.R
\name{phy_add_metadata_variables}
\alias{phy_add_metadata_variables}
\title{Add metadata variables to a phyloseq}
\usage{
phy_add_metadata_variables(physeq, df, by)
}
\arguments{
\item{physeq}{A `phyloseq` object}

\item{df}{A `data.frame` object, with rownames matching `sample_names` of the `phyloseq` object}

\item{by}{A `character` containing the variable to merge the two `sample_data` by.}
}
\description{
Adds a set of metadata variables to the metadata of a phyloseq object. Make sure `rownames` contain the same names. Row ordering is taken care of internally
}
\examples{
data(GlobalPatterns)
df_test <- data.frame(variable_new = rnorm(n = nsamples(GlobalPatterns), mean = 2, sd = 0.8))
rownames(df) <- samples_names(GlobalPatterns)
phy_add_metadata_variables(physeq=GlobalPatterns,
                           df = df_test)

}
