% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOTFINISHED_barplotTaxa.R
\name{phy_barplotTaxa}
\alias{phy_barplotTaxa}
\title{Boxplot of a taxon}
\usage{
phy_barplotTaxa(
  physeq,
  x = "Sample",
  y = "Abundance",
  fill = NULL,
  title = NULL,
  facet_grid = NULL,
  border_color = NA
)
}
\arguments{
\item{border_color}{}
}
\value{

}
\description{
Boxplot of a taxon
}
