% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy_BasicStats.R
\name{phy_BasicStats}
\alias{phy_BasicStats}
\title{Basic Phyloseq stats}
\usage{
phy_BasicStats(x, transform)
}
\arguments{
\item{x}{Either a \code{phyloseq} or a  \code{otu.table} or a \code{matrix} object. If you have a matrix, make sure taxa are on the rows}

\item{transform}{Any transformation function allowed by phy_transform}
}
\value{
A \code{data.frame} object
}
\description{
Generate a data frame with prevalence, mean abundance and variance of the taxa in a phyloseq object
}
\examples{

data(GlobalPatterns)
gp_genus <- tax_glom(GlobalPatterns, "Genus")
basic_statistics <- phy_BasicStats(gp_genus, transform = "clr")

}
