% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy_betaPlotTaxon.R
\name{phy_betaPlotTaxon}
\alias{phy_betaPlotTaxon}
\title{Fancy MDS Plots}
\usage{
phy_betaPlotTaxon(physeq, dist = "bray", method = "PCoA", axes = 1:2, taxon)
}
\arguments{
\item{physeq}{A phyloseq object, **CAREFUL**: the function doesn't transform internally, so if you choose to calculate the distance matrix internally with "ordinate", make sure you transform the counts prior to function call (usually relative abundance is accepted)}

\item{dist}{distance object or character vector saying the distance metric to apply}

\item{method}{one of the methods allowed by "ordinate" in the phyloseq object}

\item{axes}{The axes to be plotted, default are the first 2, which should discriminate samples better}

\item{taxon}{the taxon name to color the samples by. The palette is "viridis_c"}
}
\value{
a ggplot2 object
}
\description{
The function wraps the phyloseq functions "distance", "ordinate", and "plot_ordination".
This function is useful to see the gradient of abundance of a taxon in the 2D ordinated dataset
}
\examples{
data("enterotype")

phy_betaPlotTaxon(physeq = enterotype,
dist = "bray",
method = "PCoA",
axes = 1:2,
taxon = "Bacteroides")
}
