% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy_drop_NA_vars.R
\name{phy_drop_NA_vars}
\alias{phy_drop_NA_vars}
\title{Drop NAs from selected variables}
\usage{
phy_drop_NA_vars(physeq, vars)
}
\arguments{
\item{physeq}{a phyloseq object}

\item{vars}{a character vector with variable names to remove NAs from}
}
\value{
a clean phyloseq object
}
\description{
Given a phyloseq object and a character vector, the function removes all NAs that are in at least one variable
}
\examples{
data("GlobalPatterns")
set.seed(1111)
GlobalPatternsNA <- GlobalPatterns
GlobalPatternsNA@sam_data$newVar1 <- replicate(n = nsamples(GlobalPatternsNA), expr = sample(x =c(NA, 1,0), size = 1, prob = c(0.1, 0.4, 0.5)))
GlobalPatternsNA@sam_data$newVar2 <- replicate(n = nsamples(GlobalPatternsNA), expr = sample(x =c(NA, 1,0), size = 1, prob = c(0.1, 0.4, 0.5)))

drop_NA_from_phyloseq_vars(physeq = GlobalPatternsNA, vars = c("newVar1", "newVar2"))
GlobalPatterns
}
