% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy_heatmap.R
\name{phy_heatmap}
\alias{phy_heatmap}
\title{Heatmap of microbiota features}
\usage{
phy_heatmap(
  physeq,
  variable,
  physeq_transformation = "identity",
  taxa = "all",
  aggregation_samples = "median",
  taxa_into_rows = TRUE,
  scale = "row",
  add_sample_size = TRUE,
  angle_col = "90",
  legend = FALSE,
  treeheight_row = 20,
  treeheight_col = 20
)
}
\arguments{
\item{physeq}{A phyloseq object}

\item{variable}{The variable to consider for clustering}

\item{physeq_transformation}{The transformation you want to apply to your data. Default is \code{identity}, which is no transformation}

\item{taxa}{A facultative list of taxa to subset the phyloseq by. Default is \code{all}}

\item{aggregation_samples}{If you want to perform variable-wise aggregation. the allowed aggregation methods here are \code{mean, median, none}. NB: \code{none} doesn't work yet}

\item{taxa_into_rows}{A \code{logical}. It asks whether you want taxa as rows or columns of the heatmap. make sure you do the scaling right, then, in the next parameter}

\item{scale}{The \code{scale} option in \code{pheatmap}, either scale by row or column. default is "row", as also \code{taxa_into_rows} default is TRUE}

\item{add_sample_size}{A \code{logical}. It asks whether you want to add the per-category sample size in your variable's labels}

\item{angle_col}{A \code{numeric/character} value of the degrees of tilting of labels. Justification is taken care of automatically}

\item{legend}{A \code{logical}. It asks whether you want to draw the legend or not}

\item{treeheight_row}{A \code{numeric} value for drawing the height of the row-wise tree clustering}

\item{treeheight_col}{A \code{numeric} value for drawing the height of the column-wise tree clustering}
}
\description{
Make a quick heatmap of selected features, also transforming internally
}
\examples{
data(enterotype)

phy_heatmap(physeq = enterotype,
            variable = "SeqTech",
            aggregation_samples = "median",
            taxa_into_rows = TRUE,
            scale = "row",
            add_sample_size = TRUE,
            angle_col = 45,
            legend = FALSE)

}
