% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy_OtuMetaTable.R
\name{phy_OtuMetaTable}
\alias{phy_OtuMetaTable}
\title{merge otu and metadata table}
\usage{
phy_OtuMetaTable(physeq)
}
\arguments{
\item{physeq}{a phyloseq object with otu and metadata tables available}
}
\value{
a data.frame object, with otus as first columns, and metadata merged after them
}
\description{
merge otu and metadata table
}
\examples{

data(GlobalPatterns)0

otuMeta <- OtuMetaTable(GlobalPatterns)
}
