% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy_plotTaxon.R
\name{phy_plotTaxon}
\alias{phy_plotTaxon}
\title{QuickPlot of a taxon}
\usage{
phy_plotTaxon(
  physeq,
  taxon,
  trait,
  colour = NULL,
  transform = "clr",
  jitter_alpha = NULL
)
}
\arguments{
\item{physeq}{a phyloseq object}

\item{taxon}{the name of the taxon you want to plot (y axis)}

\item{trait}{the name of the trait you want on your x axis}

\item{colour}{another variable you may want to color your points by}

\item{transform}{a transformation, all those in the phy_transform are allowed}

\item{jitter_alpha}{logical for whether you want to see the dots or not (careful, time-consuming step)}

\item{fill}{another variable you may want to fill your non-points by}
}
\value{
a ggplot2 object
}
\description{
QuickPlot of a taxon
}
\examples{


}
