% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy_plotTaxon_boxplot.R
\name{phy_plotTaxon_boxplot}
\alias{phy_plotTaxon_boxplot}
\title{Boxplot of a taxon}
\usage{
phy_plotTaxon_boxplot(physeq, y, x, fill, transform = NULL)
}
\arguments{
\item{physeq}{a phyloseq object}

\item{y}{the name of the taxon you want to plot (y axis)}

\item{x}{the name of the trait you want to plot (x axis)}

\item{fill}{another variable you may want to fill your boxes by}

\item{transform}{a transformation, all those in the phy_transform are allowed}
}
\value{
a ggplot2 object
}
\description{
Boxplot of a taxon
}
\examples{
data("enterotype")

phy_plotTaxon_boxplot(physeq = enterotype,
                     y = "Rikenella",
                     x = "Enterotype",
                     fill = "SeqTech",
                     transform = "clr")
                    
}
