% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy_plotTaxon_hist.R
\name{phy_plotTaxon_hist}
\alias{phy_plotTaxon_hist}
\title{Quick histogram of taxa}
\usage{
phy_plotTaxon_hist(
  physeq,
  taxa,
  color_or_fill,
  transform = "clr",
  alpha = 1,
  geom = "histogram"
)
}
\arguments{
\item{physeq}{a phyloseq object}

\item{taxa}{a character, the ID(s) of the taxa to plot}

\item{transform}{a character to transform the taxa with, any from "phy_transform" are allowed}

\item{alpha}{a double, transparency of the plot elements: between 0 (completely transparent) and 1 (full density)}

\item{geom}{character, either "histogram" or "density"}

\item{color_fill}{a character, the variable to color or fill the histogram or the density plot by}
}
\value{
a ggplot2 object
}
\description{
Useful if you need a quick histogram or density plot of 1 or more taxa, which you can also color by another trait
}
\examples{

data("GlobalPatterns")

# fix a bug in the ggplot2 code, which does not accept numeric variables as string names
taxa_names(GlobalPatterns) <- paste0("t",taxa_names(GlobalPatterns))

phy_plotTaxon_hist(physeq = GlobalPatterns, taxa = "t549656", color_fill = "SampleType", geom="density")
phy_plotTaxon_hist(physeq = GlobalPatterns, taxa = "t549656", color_fill = "SampleType", geom="histogram")

}
