% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy_plotTaxon_scatter.R
\name{phy_plotTaxon_scatter}
\alias{phy_plotTaxon_scatter}
\title{Scatterplot of a taxon}
\usage{
phy_plotTaxon_scatter(physeq, y, x, colour = NULL, transform = NULL)
}
\arguments{
\item{physeq}{a phyloseq object}

\item{y}{the name of the taxon you want to plot (y axis)}

\item{x}{the name of the trait you want to plot (x axis)}

\item{colour}{another variable you may want to color your points by}

\item{transform}{a transformation, all those in the phy_transform are allowed}
}
\value{
a ggplot2 object
}
\description{
Make sure the "x" parameter calls a continuous variable
}
\examples{
data("enterotype")
enterotype@sam_data$otherVar <- rnorm(n = nsamples(enterotype), mean = 5, sd = 1.5)

phy_plotTaxon_scatter(physeq = enterotype,
                     y = "Rikenella",
                     x = "otherVar",
                     colour = "SeqTech",
                    transform = "clr")

}
