% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy_substitute_metadata.R
\name{phy_substitute_metadata}
\alias{phy_substitute_metadata}
\title{Substitute a phyloseq metadata with a new data.frame}
\usage{
phy_substitute_metadata(physeq, new_metadata)
}
\arguments{
\item{physeq}{A phyloseq object contaning or not containing `sample_data`}

\item{new_metadata}{a data.frame object, with rownames containing `sample_names` of the `physeq` object}
}
\value{
a phyloseq object with the new metadata stored under the "sam_data" slot
}
\description{
Substitute metadata in phyloseq objects. the two data frames must have row names matching sample names. The sample order doesn't matter, because the functions will reorder the new metatadata rows, to match the sample order in the phyloseq object.
}
\examples{
data("GlobalPatterns")
exMeta <- microbiome::meta(GlobalPatterns)
newMeta <- exMeta \%>\% mutate(newVar = runif(n = nrow(exMeta), min =  0, max = 20))

GlobalPatterns <- phy_substitute_metadata(physeq = GlobalPatterns, new_metadata = newMeta)

}
