% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy_transform.R
\name{phy_transform}
\alias{phy_transform}
\title{Transform microbiota counts}
\usage{
phy_transform(physeq, transform)
}
\arguments{
\item{physeq}{a phyloseq object}

\item{transform}{all options available in "microbiome::transform()", extending it with the "asinh" option}
}
\value{

}
\description{
Same as microbiome::transform(), but also adding the hyperbolic arcsine
}
\examples{
data(GlobalPatterns)
GP_Genus <- tax_glom(GlobalPatterns, "Genus")
tax <- "295395"
transf <- "clr"
transformed_GP <- transform_microbiota_ga(GP_Genus,transform = transf)


plot(density(microbiome::abundances(GP_Genus)[tax,]), main = paste("taxon n°", tax), col = "blue")
lines(density(microbiome::abundances(transformed_GP)[tax,]), col = "red")
legend("topright", legend = c("Untransformed", transf), col = c("blue", "red"), pch = 18)

}
