% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_to_fasta.R
\name{phyloseq_to_fasta}
\alias{phyloseq_to_fasta}
\title{Title}
\usage{
phyloseq_to_fasta(physeq, destination_file = NULL)
}
\arguments{
\item{physeq}{a phyloseq object}

\item{destination_file}{write the file you want your fasta sequences to be written into}
}
\value{
either a list of sequences or it writes them into the specified file
}
\description{
Title
}
\examples{

}
