% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickEvenSamples_factor.R
\name{pickEvenSamples_factor}
\alias{pickEvenSamples_factor}
\title{Pick samples proportionally from factor}
\usage{
pickEvenSamples_factor(x, factor, proportion, seed = NULL)
}
\arguments{
\item{x}{a \code{phyloseq} or \code{data.frame} object}

\item{factor}{a \code{character} vector with the name of the non-numeric variable to split the data set by}

\item{proportion}{the proportion of samples you want to get from each category in the factor above}

\item{seed}{if you want to specify your seed to ensure perfect replication, feel free}
}
\value{
A \code{list} containing the subsetted data and the remaining data, which you can use as you wish. I thought them as a "training" and "test" subset
}
\description{
A very simple function that helps picking some samples. I thought it to split data into a training and test set for machine learning puroposes, but it is not perfect. 
The next step would be to allow some other covariates to be evenly distributed across the subsets.
}
