% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qklook.R
\name{qklook}
\alias{qklook}
\title{Glimpse into a table-like object
A simple function to make a "head" of the data on both rows and column, of custom length}
\usage{
qklook(x, nrows = 8, ncols = 8)
}
\arguments{
\item{x}{H \code{data.frame} or \code{matrix-array} object}

\item{nrows}{How many rows to print (default = 8)}

\item{ncols}{How many columns to print (default = 8)}
}
\value{
a printed subset of the data
}
\description{
Glimpse into a table-like object
A simple function to make a "head" of the data on both rows and column, of custom length
}
\examples{
data(iris)
qklook(iris)
qklook(iris, 5,5)
}
