% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samp_cond_medians_IQR.R
\name{samp_cond_medians_IQR}
\alias{samp_cond_medians_IQR}
\title{Create Sample conditional means and standard deviation}
\usage{
samp_cond_medians_IQR(x, g, na.rm = TRUE)
}
\arguments{
\item{x}{the numeric/integer/double vector of values}

\item{g}{the factor or character vector to group values by}

\item{na.rm}{should}
}
\value{
a matrix object, containing median and IQR on two columns, and group names in rows
}
\description{
Create Sample conditional means and standard deviation
}
\examples{

data("gss_cat")
age_by_marital_status <- samp_cond_medians_IQR(gss_cat$age, g = gss_cat$marital)
age_by_marital_status

}
