% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_palette_colors.R
\name{show_palette_colors}
\alias{show_palette_colors}
\title{View the colors of a HTML list}
\usage{
show_palette_colors(html_codes)
}
\arguments{
\item{html_codes}{HTML codes you want to have a look at}
}
\value{
A figure showing the palette colors
}
\description{
View the colors of a HTML list
}
\examples{
colors_vec <- ggsci::pal_futurama()(4)

show_palette_colors(colors_vec)

}
