% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaPlot.R
\name{betaPlot}
\alias{betaPlot}
\title{Fancy MDS Plots}
\usage{
betaPlot(
  physeq,
  dist = "bray",
  method = "PCoA",
  axes = 1:2,
  color,
  shape = color,
  visual_grouping = NULL,
  palette = "standard"
)
}
\arguments{
\item{physeq}{A phyloseq object, **CAREFUL**: the function doesn't transform internally, so if you choose to calculate the distance matrix internally with "ordinate", make sure you transform the counts prior to function call (usually relative abundance is accepted)}

\item{dist}{Distance object or character vector saying the distance metric to apply}

\item{method}{One of the methods allowed by "ordinate" in the phyloseq object}

\item{axes}{The axes to be plotted, default are the first 2, which should discriminate samples better}

\item{color}{The variable to color the samples by. Only metadata variables are supported, to color taxa, use betaPlotTaxa}

\item{shape}{Should you want to further differentiate your points. the default is that shapes are chosen based on the color variable}

\item{visual_grouping}{This is the cool parameter of this function. the allowed ones are "spiders", "hulls" and "ellipses"}

\item{palette}{Similarly to ggpubr's "palette" parameters, the function allows for a palette name, (default is "standard", the ggplot2 default palette). All palettes in the ggsci package are allowed too}
}
\value{
a ggplot2 object
}
\description{
The function wraps the phyloseq functions "distance", "ordinate", and "plot_ordination",
also adding some fancier represantations, called "spiders", "hulls" and "ellipses".
}
\examples{
data("enterotype")

betaPlot(physeq = enterotype,
dist = "bray",
method = "PCoA",
axes = 1:2,
color = "SeqTech",
visual_grouping = "spiders",
palette = "jco")
}
