% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_demogr_table.R
\name{make_demogr_table}
\alias{make_demogr_table}
\title{Make a complex demographic table}
\usage{
make_demogr_table(
  df,
  row_variables_as_are,
  row_var_names_nicer,
  col_variable,
  col_var_names_nicer,
  add_perc_column,
  caption = NULL
)
}
\arguments{
\item{df}{A `data.frame` or `tibble` containing several variables of interest to make a table from}

\item{row_variables_as_are}{a `vector` of column names to subdivide rows by}

\item{row_var_names_nicer}{(Facultative) a character vector of variables names in `row_variables_as_are`, but more "presentation ready"}

\item{col_variable}{A `character` vector with the name of the df colum which will become columns of the}

\item{col_var_names_nicer}{if you want nicer-looking names to the column variable items/levels,}

\item{add_perc_column}{Add per-table-column's percentage column}

\item{caption}{. Default is `NULL`}
}
\value{

}
\description{
Summarize multiple **categorical** variable columns in one handy function
NB: Doesn't yet work perfectly, also, it doesn't make summary statistics of continuous variables
}
\examples{
make_demogr_table(df= mtcars,
row_variables_as_are = c("mpg","gear", "carb"),
row_var_names_nicer = c("Miles per gallon", "N° gears", "Fuel type"),
col_variable = "cyl",
add_perc_column = TRUE)

}
