% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaphlan_to_phyloseq.R
\name{metaphlan_to_phyloseq}
\alias{metaphlan_to_phyloseq}
\title{Read Metaphlan profiles}
\usage{
metaphlan_to_phyloseq(mpa_path, metadata = NULL, phyloTree = NULL, version = 4)
}
\arguments{
\item{version}{the metaphlan version, by default, version 4 is taken (to date it is the latest - Apr 2022)}

\item{path}{a \code{character} vector containing the path to the text table to read}
}
\value{

}
\description{
Given the table location, the function
}
