% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phy_barplotTaxa.R
\name{phy_barplotTaxa}
\alias{phy_barplotTaxa}
\title{Barplot taxa (beta)}
\usage{
phy_barplotTaxa(
  physeq,
  x = "Sample",
  tax_fill = NULL,
  title = NULL,
  facet = NULL,
  transform = "compositional"
)
}
\arguments{
\item{physeq}{A phyloseq object}

\item{x}{The x axis variable, default is "Sample", which will plot one column for each sample}

\item{tax_fill}{The taxonomic level to fill bars by}

\item{title}{A character vector to entitle the plot with}

\item{facet}{A character vector of a metadata variable's name}

\item{transform}{A character vector of the transformation method to apply}
}
\value{

}
\description{
A not-fully-finished but "MVP" (minimum viable product) function that plots phyloseq data preferrably in relative abundance.
}
\examples{

data("GlobalPatterns")

testplotSample <- phy_barplotTaxa(physeq = GlobalPatterns,
                                 x = "Sample",
                                 tax_fill = "Genus",
                                 title = "By Sample",
                                 facet = "SampleType",
                                 transform = "compositional")

testplotSample <- phy_barplotTaxa(physeq = GlobalPatterns,
                                 x = "SampleType",
                               tax_fill = "Genus",
                               title = "By Sample",
                               transform = "compositional")

}
